fix: split generic-call inference helpers to satisfy file-size ratchet
Relocate position_is_regeneralizable_higher_order_target and
single_concrete_upper_bound from inference_helpers.rs into a focused
sibling module inference_helpers_higher_order.rs (pure relocation,
behavior-preserving). Brings inference_helpers.rs from 2094 to 1999
lines, under the 2065 ratchet limit.
