+++ crates/tsz-solver/src/operations/generic_call/mod.rs
@@ -661,6 +661,7 @@ mod contextual_signature_instantiation;
 mod foreign_param_shapes;
 mod inference_helpers;
 mod inference_helpers_candidates;
+mod inference_helpers_higher_order;
 mod normalization;
 mod readonly_direct_inference;
 pub mod request;
